@@ -11,28 +11,27 @@ using Microsoft.AspNetCore.Mvc;
 using System.Net.Http;
 using Microsoft.Extensions.Configuration;
 
-namespace BotoGP.Web.Controllers
+namespace BotoGP.Web.Controllers;
+
+[Route("api/[controller]")]
+public class EnvironmentController : Controller
 {
-	[Route("api/[controller]")]
-    public class EnvironmentController : Controller
-    {
-        public IConfiguration Config { get; }
+	public IConfiguration Config { get; }
 
-        public EnvironmentController(IConfiguration config)
-        {
-            Config = config;
-        }
+	public EnvironmentController(IConfiguration config)
+	{
+		Config = config;
+	}
 
 
-		// GET: api/values
-		[HttpGet]
-		public IEnumerable<string> Get()
+	// GET: api/values
+	[HttpGet]
+	public IEnumerable<string> Get()
+	{
+		return new[]
 		{
-            return new[]
-            {
-                "Config keys: " + string.Join(", ", Config.AsEnumerable().Select(x => x.Key)),
-                "AppSettings keys: " + string.Join(", ", Config.GetSection("AppSettings").AsEnumerable().Select(x => x.Key))
-            };
-		}
-    }
+			"Config keys: " + string.Join(", ", Config.AsEnumerable().Select(x => x.Key)),
+			"AppSettings keys: " + string.Join(", ", Config.GetSection("AppSettings").AsEnumerable().Select(x => x.Key))
+		};
+	}
 }
