@@ -1,13 +1,12 @@
 ﻿
 using Microsoft.Extensions.Configuration;
 
-namespace BotoGP.Domain.Helpers
+namespace BotoGP.Domain.Helpers;
+
+public static class ConfigReader
 {
-    public static class ConfigReader
+    public static string ReadAppSetting(IConfiguration configuration, string name)
     {
-        public static string ReadAppSetting(IConfiguration configuration, string name)
-        {
-            return configuration["APPSETTING_" + name] ?? configuration["AppSettings:" + name];
-        }
+        return configuration["APPSETTING_" + name] ?? configuration["AppSettings:" + name];
     }
 }
