@@ -1,59 +1,58 @@
 ﻿
 
-namespace BotoGP.Domain.Services
+namespace BotoGP.Domain.Services;
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Threading.Tasks;
+using BotoGP.stateserver.Models;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json.Linq;
+
+public class CircuitRepository : ICircuitRepository
 {
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Net.Http;
-    using System.Threading.Tasks;
-    using BotoGP.stateserver.Models;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.Logging;
-    using Newtonsoft.Json.Linq;
+    IConfiguration _configuration;
 
-    public class CircuitRepository : ICircuitRepository
+    ILogger _logger;
+
+    private static HttpClient client = new HttpClient();
+
+    public CircuitRepository(IConfiguration configuration, ILogger<CircuitRepository> logger)
     {
-        IConfiguration _configuration;
-
-        ILogger _logger;
-
-        private static HttpClient client = new HttpClient();
-
-        public CircuitRepository(IConfiguration configuration, ILogger<CircuitRepository> logger)
-        {
-            _configuration = configuration;
-            _logger = logger;
+        _configuration = configuration;
+        _logger = logger;
             
-            _logger.LogInformation("StorageConnectionString");
-            _logger.LogInformation(_configuration["StorageConnectionString"]);
-        }
+        _logger.LogInformation("StorageConnectionString");
+        _logger.LogInformation(_configuration["StorageConnectionString"]);
+    }
 
-        public async Task<IEnumerable<Circuit>> ReadAll()
-        {
-            var result = await client.GetStringAsync(buildUrl("/circuits"));
+    public async Task<IEnumerable<Circuit>> ReadAll()
+    {
+        var result = await client.GetStringAsync(buildUrl("/circuits"));
 
-            var array = Newtonsoft.Json.JsonConvert.DeserializeObject(result) as JArray;
+        var array = Newtonsoft.Json.JsonConvert.DeserializeObject(result) as JArray;
 
-            return array.ToObject<List<Circuit>>();
-        }
+        return array.ToObject<List<Circuit>>();
+    }
 
-        public async Task<Circuit> Read(string id)
-        {
-            var result = await client.GetStringAsync(buildUrl("/circuit?id=" + id));
+    public async Task<Circuit> Read(string id)
+    {
+        var result = await client.GetStringAsync(buildUrl("/circuit?id=" + id));
 
-            return Newtonsoft.Json.JsonConvert.DeserializeObject(result) as Circuit;
-        }
+        return Newtonsoft.Json.JsonConvert.DeserializeObject(result) as Circuit;
+    }
 
-        public async Task Store(Circuit circuit)
-        {
-            var req = new HttpRequestMessage(HttpMethod.Post, buildUrl("/circuits?key=" + circuit.Id.ToString()));
-            req.Content = new StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(circuit), System.Text.Encoding.UTF8, "application/json");
-            await client.SendAsync(req);
-        }
+    public async Task Store(Circuit circuit)
+    {
+        var req = new HttpRequestMessage(HttpMethod.Post, buildUrl("/circuits?key=" + circuit.Id.ToString()));
+        req.Content = new StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(circuit), System.Text.Encoding.UTF8, "application/json");
+        await client.SendAsync(req);
+    }
 
-        string buildUrl(string path)
-        {
-            return Helpers.ConfigReader.ReadAppSetting(_configuration, "CircuitsRepositoryBaseUrl") + path;
-        }
+    string buildUrl(string path)
+    {
+        return Helpers.ConfigReader.ReadAppSetting(_configuration, "CircuitsRepositoryBaseUrl") + path;
     }
 }
