@@ -2,28 +2,26 @@
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace BotoGP.stateserver.Models
+namespace BotoGP.stateserver.Models;
+
+public class Circuit
 {
-    public class Circuit
-    {
-        public Guid Id { get; set; } = Guid.NewGuid();
+	public Guid Id { get; set; } = Guid.NewGuid();
 
-        public string Name { get; set; } = "";
+	public string Name { get; set; } = "";
 
-        public int Width { get; set; } = 150;
+	public int Width { get; set; } = 150;
 
-		public int Height { get; set; } = 100;
+	public int Height { get; set; } = 100;
 
-        public int Scale { get; set; } = 5;
+	public int Scale { get; set; } = 5;
 
-        public string Checkpoints { get; set; }
+	public string Checkpoints { get; set; }
         
-        public CircuitMap Map { get; set; } = new CircuitMap();
+	public CircuitMap Map { get; set; } = new CircuitMap();
 
-		public override string ToString()
-		{
-			return JsonConvert.SerializeObject(this);
-		}
-    }
+	public override string ToString()
+	{
+		return JsonConvert.SerializeObject(this);
+	}
 }
- 
