@@ -2,20 +2,19 @@
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace BotoGP.stateserver.Models
+namespace BotoGP.stateserver.Models;
+
+public class CircuitMap
 {
-    public class CircuitMap
+    [JsonProperty(PropertyName = "checkPoints")]
+    public List<CheckPoint> CheckPoints {get;set;} = new List<CheckPoint>();
+
+    public List<CheckPoint> OnTrack { get; set; } = new List<CheckPoint>();
+
+    public List<CheckPoint> OffTrack { get; set; } = new List<CheckPoint>();
+
+    public override string ToString()
     {
-        [JsonProperty(PropertyName = "checkPoints")]
-		public List<CheckPoint> CheckPoints {get;set;} = new List<CheckPoint>();
-
-        public List<CheckPoint> OnTrack { get; set; } = new List<CheckPoint>();
-
-        public List<CheckPoint> OffTrack { get; set; } = new List<CheckPoint>();
-
-        public override string ToString()
-        {
-            return JsonConvert.SerializeObject(this);
-        }
+        return JsonConvert.SerializeObject(this);
     }
 }
