@@ -1,16 +1,15 @@
 using System;
 using System.Collections.Generic;
 
-namespace BotoGP.stateserver.Models
+namespace BotoGP.stateserver.Models;
+
+public class UpdateCircuitDto
 {
-    public class UpdateCircuitDto
-    {
-        public string Name { get; set; }
+    public string Name { get; set; }
         
-        public List<CheckPoint> CheckPoints { get; set; }
+    public List<CheckPoint> CheckPoints { get; set; }
 
-        public List<CheckPoint> OnTrack { get; set; }
+    public List<CheckPoint> OnTrack { get; set; }
 
-        public List<CheckPoint> OffTrack { get; set; }
-    }
+    public List<CheckPoint> OffTrack { get; set; }
 }
