@@ -4,15 +4,14 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace BotoGP.Web.Pages
-{
-    public class AboutModel : PageModel
-    {
-        public string Message { get; set; }
+namespace BotoGP.Web.Pages;
 
-        public void OnGet()
-        {
-            Message = "Your application description page.";
-        }
+public class AboutModel : PageModel
+{
+    public string Message { get; set; }
+
+    public void OnGet()
+    {
+        Message = "Your application description page.";
     }
 }
