@@ -8,19 +8,18 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace BotoGP.Web
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            BuildWebHost(args).Run();
-        }
+namespace BotoGP.Web;
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseWebRoot(Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "public"))
-                .UseStartup<Startup>()
-                .Build();
+public class Program
+{
+    public static void Main(string[] args)
+    {
+        BuildWebHost(args).Run();
     }
+
+    public static IWebHost BuildWebHost(string[] args) =>
+        WebHost.CreateDefaultBuilder(args)
+            .UseWebRoot(Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "public"))
+            .UseStartup<Startup>()
+            .Build();
 }
