@@ -56,7 +56,7 @@
         <span class="sr-only">Previous</span>
     </a>
     <a class="right carousel-control" href="#circuit-carousel" role="button" data-slide="next">
-        <span class="icon-next aria-hidden="true"></span>
+        <span class="icon-next" aria-hidden="true"></span>
         <span class="sr-only">Next</span>
     </a>
 </div>
